Add dynamic speaking rate control demo. Fix SinkAttention bug

+++ voxtream/dataset.py
@@ -212,7 +212,9 @@ class TrainDataset(Dataset):
         mask = (ins_idx > min_idx) & (ins_idx <= max_idx + 1)
         ins_in_slice = ins_idx[mask] - min_idx
         phoneme_sequence = np.insert(phoneme_sequence, ins_in_slice, tokens[mask])
-        punct_del_idx = (ins_in_slice + np.arange(mask.sum())).astype(self.dtype) + 1  # <BOS>
+        punct_del_idx = (ins_in_slice + np.arange(mask.sum())).astype(
+            self.dtype
+        ) + 1  # <BOS>
 
         # Add prompt <UNK> phone tokens
         phoneme_sequence = np.concatenate(
